apps/api/requirements.txt
flask==3.0.3
flask-cors==4.0.1
strawberry-graphql==0.246.2
psycopg[binary]==3.2.3
